utils: Moved malloc/realloc into sep. file
